# Avoid memoizing objects with 1 refcount