If an object has a refcount of 1, it doesn't need to be memoized. For
messages where this optimization applies, this dramatically reduces the
cost of enabling memoization (~3x faster), and reduces the size of the
serialized message.

One complication here is tuples. Since tuples are supposed to be
immutable, cycles are detected through memoization of the tuples
elements, not the tuple itself. Since a cyclic tuple would have > 1
reference, but the elements might not, we need to add some extra logic
to ensure that in the presence of cyclic tuples the elements are still
memoized. This could be alleviated through new opcodes (since quickle
doesn't ever pass tuples to other callables/objects, mutating tuples is
fine here), but that would further break with pickle and might not be
necessary. Maybe something to investigate later.

Also fixes (and adds tests for) a few refcounting bugs in the struct
class methods. References weren't being properly decrefed before.