@jcrist 2020-08-30
A quick benchmark:

```python
import pickle
import quickle
import uuid

def record():
    data = {str(uuid.uuid4().hex): str(uuid.uuid4().hex) for _ in range(4)}
    data[str(uuid.uuid4().hex)] = [1, 2, 3, 4, 5]
    return data

data = [record() for _ in range(10000)]

enc_memoize = quickle.Encoder(memoize=True)
enc = quickle.Encoder(memoize=False)
```

**Master**

```python
In [1]: %timeit enc.dumps(data)
4.53 ms ± 21.5 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [2]: %timeit enc_memoize.dumps(data)
11.2 ms ± 48.8 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

**This PR**

```python
In [1]: %timeit enc.dumps(data)
4.52 ms ± 21 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [2]: %timeit enc_memoize.dumps(data)
4.63 ms ± 11.5 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

So no measurable slowdown from the added check, and the speedups when memoization is enabled resulted in a ~2.4x speedup.